# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Savne Saltcorn Agent API",
    "version": "0.2.55",
    "description": "Authenticated administrative API for operating Saltcorn safely from agents, scripts, and automation tools. All endpoints are served under /savne-saltcorn-agent-api."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Saltcorn instance"
    }
  ],
  "security": [
    {
      "SaltcornUserApiToken": []
    }
  ],
...
  "paths": {
...
    "/savne-saltcorn-agent-api/commands": {
      "get": {
        "tags": ["Discovery"],
        "summary": "List supported commands",
        "operationId": "commands",
        "responses": {
          "200": {
            "description": "Supported commands",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CommandsResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/state": {
      "get": {
        "tags": ["Discovery"],
        "summary": "Inspect visible Saltcorn state",
        "operationId": "state",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/validate": {
      "post": {
        "tags": ["Discovery"],
        "summary": "Validate the access layer",
        "operationId": "validate",
        "requestBody": { "$ref": "#/components/requestBodies/AnyJson" },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/viewtemplates": {
      "get": {
        "tags": ["Discovery"],
        "summary": "List available Saltcorn viewtemplates",
        "operationId": "viewtemplates",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/fields/capabilities": {
      "get": {
        "tags": ["Discovery"],
        "summary": "List field types, fieldviews, and supported presets",
        "operationId": "fieldCapabilities",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/capabilities": {
      "get": {
        "tags": ["Discovery"],
        "summary": "List installed plugins and feature recipe availability",
        "operationId": "capabilities",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/actions": {
      "get": {
        "tags": ["Automation"],
        "summary": "List Saltcorn actions, trigger events, and existing triggers",
        "description": "Use this endpoint before creating triggers or webhooks. Action names and event names can vary by Saltcorn version and installed plugins.",
        "operationId": "listActionsAndTriggers",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/capabilities/require": {
      "post": {
        "tags": ["Discovery"],
        "summary": "Require one or more optional feature recipes",
        "operationId": "requireCapabilities",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RequireCapabilitiesRequest" },
              "examples": {
                "advancedViews": {
                  "value": {
                    "features": ["tabulator", "kanban", "calendar", "chart", "attachments", "many_to_many"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/refresh": {
      "post": {
        "tags": ["Maintenance"],
        "summary": "Refresh Saltcorn caches and notify workers",
        "operationId": "refreshCaches",
        "requestBody": { "$ref": "#/components/requestBodies/AnyJson" },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/settings/locale": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get Saltcorn locale settings",
        "operationId": "getLocaleSettings",
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Settings"],
        "summary": "Set Saltcorn default locale",
        "operationId": "setLocaleSettings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LocaleRequest" },
              "examples": {
                "setSpanish": {
                  "value": { "locale": "es", "dry_run": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/objects/delete": {
      "post": {
        "tags": ["Maintenance"],
        "summary": "Delete explicitly named objects only",
        "description": "Deletes only explicit page, view, and table names. Wildcards, broad cleanup, and the users table are rejected.",
        "operationId": "deleteObjects",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteObjectsRequest" },
              "examples": {
                "dryRun": {
                  "value": {
                    "pages": ["productos"],
                    "views": ["Grid operativo de productos"],
                    "tables": ["productos"],
                    "dry_run": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/tables": {
      "post": {
        "tags": ["Schema"],
        "summary": "Create a table and optional fields",
        "operationId": "createTable",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTableRequest" },
              "examples": {
                "productos": {
                  "value": {
                    "name": "productos",
                    "label": "Productos",
                    "fields": [
                      { "name": "codigo", "label": "Codigo", "type": "String", "required": true, "unique": true },
                      { "name": "nombre", "label": "Nombre", "type": "String", "required": true },
                      { "name": "activo", "label": "Activo", "type": "Bool" }
                    ],
                    "dry_run": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/fields": {
      "post": {
        "tags": ["Schema"],
        "summary": "Add a field to an existing table",
        "operationId": "addField",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddFieldRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/relations/key": {
      "post": {
        "tags": ["Schema"],
        "summary": "Create a many-to-one Key relation",
        "operationId": "createKeyRelation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateKeyRelationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/relations/many-to-many": {
      "post": {
        "tags": ["Schema"],
        "summary": "Create or wire a many-to-many relation",
        "operationId": "createManyToManyRelation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateManyToManyRelationRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/rows": {
      "post": {
        "tags": ["Data"],
        "summary": "Insert rows into a table",
        "operationId": "createRows",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRowsRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/basic": {
      "post": {
        "tags": ["Views"],
        "summary": "Create native basic List, Show, and Edit views",
        "operationId": "createBasicViews",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBasicViewsRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/form-layout": {
      "post": {
        "tags": ["Views"],
        "summary": "Create an advanced native Edit form layout",
        "operationId": "createFormLayoutView",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateFormLayoutViewRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/tabulator": {
      "post": {
        "tags": ["Views"],
        "summary": "Create a Tabulator grid view",
        "operationId": "createTabulatorView",
        "requestBody": { "$ref": "#/components/requestBodies/ViewRecipe" },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/calendar": {
      "post": {
        "tags": ["Views"],
        "summary": "Create a FullCalendar view",
        "operationId": "createCalendarView",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCalendarViewRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/chart": {
      "post": {
        "tags": ["Views"],
        "summary": "Create a chart view",
        "operationId": "createChartView",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateChartViewRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/attachments": {
      "post": {
        "tags": ["Views"],
        "summary": "Create a Multi File Upload attachment view",
        "operationId": "createAttachmentView",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAttachmentViewRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/savne-saltcorn-agent-api/views/kanban": {
      "post": {
        "tags": ["Views"],
        "summary": "Create a Kanban board view",
        "operationId": "createKanbanView",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateKanbanViewRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/GenericOk" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SaltcornUserApiToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Saltcorn user API token",
        "description": "Native Saltcorn user API token. The user must be an admin user."
      }
    },
    "responses": {
      "GenericOk": {
        "description": "Operation result",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/GenericResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid Saltcorn user API token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Token belongs to a disabled or non-admin user",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
...
    "schemas": {
...
      "CommandsResponse": {
        "type": "object",
        "required": ["ok", "plugin", "commands"],
        "properties": {
          "ok": { "type": "boolean" },
          "plugin": { "type": "string" },
          "commands": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Command" }
          }
        },
        "additionalProperties": true
      },
...
    }
  }
}